security: Disable smartcard code by default
The smartcard code contains a number of security vulnerabilities and
does not work at the moment.

The code has been left in the source tree, but moved behind an
'--enable-smartcard' configure flag which is clearly marked as not
for production use.

(cherry picked from commit 7a2ac0c177fca194ee48392b76a1dfc67b27998e)
